Refactor RT module: Split atmrt.py into modular components
- Split 886-line atmrt.py into focused modules:
  * common.py: ArtCommon base class with shared functionality (346 lines)
  * emis.py: Emission RT classes (ArtEmisPure, ArtEmisScat) (211 lines)
  * reflect.py: Reflection RT classes (ArtReflectPure, ArtReflectEmis, ArtAbsPure) (213 lines)
  * trans.py: Transmission RT class (ArtTransPure) (118 lines)

- Implement lazy loading public API in rt/__init__.py
  * Consistent with opacity module architecture
  * Clean imports: `from exojax.rt import ArtEmisPure`
  * Tab completion support

- Maintain full backward compatibility
  * Deprecated atmrt.py with proper warnings
  * All existing imports continue to work
  * Scheduled for removal in v3.0

- Benefits:
  * 60% reduction in average file size
  * Physics-based logical organization
  * Improved maintainability and code navigation
  * Future-ready for additional RT methods

All unit tests pass (7/7 RT tests successful)
